requirements.txt
async-generator==1.10
attrs==22.2.0
autocommand==2.2.2
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
charset-normalizer==3.0.1
ConfigArgParse==1.5.3
exceptiongroup==1.1.0
filelock==3.9.0
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
jaraco.classes==3.2.3
keyring==23.13.1
mintapi==2.12
more-itertools==9.0.0
numpy==1.24.1
oathtool==2.3.1
outcome==1.2.0
pandas==1.5.2
path==16.6.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
requests-file==1.5.1
selenium==4.7.2
selenium-requests==2.0.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tldextract==3.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wsproto==1.2.0
xmltodict==0.13.0
zipp==3.11.0
